requirements.txt
gpio==0.2.0
RPi.GPIO==0.6.3
